refactor: Balance 타입을 Long으로 변경

+++ src/main/java/org/c4marathon/assignment/account/dto/request/RechargeAccountRequestDto.java
@@ -6,6 +6,6 @@ public record RechargeAccountRequestDto(
     @NotNull
     Long accountId,
     @NotNull
-    Integer balance
+    Long balance
 ) {
 }
